[KNOWAGE-6846][KNOWAGE-6848][KNOWAGE-6818][KNOWAGE-6836]
KNOWAGE-6846
Implemented both the QBE Query preview dialog, as well as the option for user to open the QBE from dataset management.
If the QBE was saved in its own dialog, dataset management detail is reloaded so the data would be consistent.

KNOWAGE-6848
This change existed in: this merge, however it got lost during some merging....
Issue is resolved and it should work as expected.

KNOWAGE-6818
Added computed property that will allow only superadmin to change category and scope. Logic is there and this condition can be easily changed.

KNOWAGE-6836
Added search as requested.